refactor villager module and move some namespaces
