+++ src/main/java/com/lothrazar/cyclicmagic/block/BlockDimensionOre.java
@@ -1,6 +1,6 @@
 package com.lothrazar.cyclicmagic.block;
 import java.util.Random;
-import com.lothrazar.cyclicmagic.registry.WorldGenModule;
+import com.lothrazar.cyclicmagic.module.WorldGenModule;
 import net.minecraft.block.BlockOre;
 import net.minecraft.block.SoundType;
 import net.minecraft.block.state.IBlockState;
