+++ src/main/java/com/lothrazar/cyclicmagic/module/VillagerCreateModule.java
@@ -1,4 +1,4 @@
-package com.lothrazar.cyclicmagic.registry;
+package com.lothrazar.cyclicmagic.module;
 import com.lothrazar.cyclicmagic.util.Const;
 import net.minecraft.entity.passive.EntityVillager;
 import net.minecraft.entity.passive.EntityVillager.EmeraldForItems;
@@ -12,10 +12,12 @@ import net.minecraftforge.fml.common.registry.VillagerRegistry;
 import net.minecraftforge.fml.common.registry.VillagerRegistry.VillagerCareer;
 import net.minecraftforge.fml.common.registry.VillagerRegistry.VillagerProfession;
 
-public class VillageTradeRegistry {
-  private static boolean extraVillagersEnabled;
-  public static void register() {
-    if (!extraVillagersEnabled) { return; }
+public class VillagerCreateModule extends BaseModule {
+  private boolean extraVillagersEnabled;
+  private void initDruid() {
+    //vanilla example :  new VillagerProfession("minecraft:butcher", "minecraft:textures/entity/villager/butcher.png");
+    String name;
+    //TO TEST: /summon Villager ~ ~ ~ {Profession:5,Career:0}
     EntityVillager.ITradeList[][] druidTrades = {
         { new EmeraldForItems(Items.COOKED_FISH, new PriceInfo(9, 12)), new EmeraldForItems(Items.APPLE, new PriceInfo(3, 6)), new EmeraldForItems(Items.BEETROOT, new PriceInfo(8, 12))
         },
@@ -28,12 +30,20 @@ public class VillageTradeRegistry {
         { new EmeraldForItems(Items.POISONOUS_POTATO, new PriceInfo(1, 3)), new EmeraldForItems(Items.SPIDER_EYE, new PriceInfo(3, 6)), new EmeraldForItems(Items.WRITTEN_BOOK, new PriceInfo(1, 1))
         }
     };
-    //		ItemStack stack = UtilNBT.buildEnchantedBook(Enchantment.getEnchantmentByID(2), (short)1);
-    //		GameRegistry.addShapelessRecipe(stack, Blocks.GRASS,Items.STICK);
-    //test to make sure the book works. if it works in craft but not in trade, then ??
+    name = "druid";
+    VillagerProfession druidProfession = new VillagerProfession(Const.MODRES + name,
+        Const.MODRES + "textures/entity/villager/" + name + ".png",
+        "minecraft:textures/entity/zombie_villager/zombie_villager.png");
+    VillagerRegistry.instance().register(druidProfession);
+    VillagerCareer druid = new VillagerCareer(druidProfession, name);
+    for (int i = 0; i < druidTrades.length; i++) {
+      druid.addTrade(i + 1, druidTrades[i]);
+    }
+  }
+  private void initSage() {
     EntityVillager.ITradeList[][] sageTrades = {
         { new EmeraldForItems(Items.GUNPOWDER, new PriceInfo(5, 8)), new EmeraldForItems(Items.NETHER_WART, new PriceInfo(12, 16))
-        //				,new ListItemForEmeralds(stack, new PriceInfo(1, 3))
+        //        ,new ListItemForEmeralds(stack, new PriceInfo(1, 3))
         },
         { new EmeraldForItems(Items.BONE, new PriceInfo(26, 32)), new EmeraldForItems(Items.MUTTON, new PriceInfo(12, 16))
         },
@@ -44,21 +54,8 @@ public class VillageTradeRegistry {
         { new EmeraldForItems(Items.GLOWSTONE_DUST, new PriceInfo(12, 16)), new ListItemForEmeralds(Items.EXPERIENCE_BOTTLE, new PriceInfo(2, 4)), new EmeraldForItems(Items.DIAMOND, new PriceInfo(1, 1)), new EmeraldForItems(Items.ENDER_PEARL, new PriceInfo(12, 16))
         }
     };
-    //vanilla example :  new VillagerProfession("minecraft:butcher", "minecraft:textures/entity/villager/butcher.png");
-    String name;
-    //TO TEST: /summon Villager ~ ~ ~ {Profession:5,Career:0}
-
-    name = "druid";
-    VillagerProfession druidProfession = new VillagerProfession(Const.MODRES + name,
-        Const.MODRES + "textures/entity/villager/" + name + ".png",
-        "minecraft:textures/entity/zombie_villager/zombie_villager.png");
-    VillagerRegistry.instance().register(druidProfession);
-    VillagerCareer druid = new VillagerCareer(druidProfession, name);
-    for (int i = 0; i < druidTrades.length; i++) {
-      druid.addTrade(i + 1, druidTrades[i]);
-    }
     //TO TEST: /summon Villager ~ ~ ~ {Profession:6,Career:0}
-    name = "sage";
+    String name = "sage";
     VillagerProfession sageProfession = new VillagerProfession(Const.MODRES + name,
         Const.MODRES + "textures/entity/villager/" + name + ".png",
         "minecraft:textures/entity/zombie_villager/zombie_villager.png");
@@ -68,7 +65,14 @@ public class VillageTradeRegistry {
       sage.addTrade(i + 1, sageTrades[i]);
     }
   }
-  public static void syncConfig(Configuration c) {
+  @Override
+  public void onInit() {
+    if (extraVillagersEnabled) {
+      initSage();
+      initDruid();
+    }
+  }
+  public void syncConfig(Configuration c) {
     String category = Const.ConfigCategory.villagers;
     c.addCustomCategoryComment(category, "Two new villagers with more trades");
     extraVillagersEnabled = c.getBoolean("More Trades", category, true, "Adds more  villager types (professions) with more trades such as gunpowder, blaze rods, beef, spider eyes, and more.  Spawn naturally and from mob eggs. ");
